Update script fix
Update script was not working, as it didn't update the git repo, because the previous compilation process left some files which are not committed. A normal git pull would fail, therefore a reset is needed to discard the changed files. Normally that's dangerous, but in our case it should be alright, as no one would be using our path to maintain that git repository.