Fixed & Improved Update Functionality (#81)
* Update script fix

Update script was not working, as it didn't update the git repo, because the previous compilation process left some files which are not committed. A normal git pull would fail, therefore a reset is needed to discard the changed files. Normally that's dangerous, but in our case it should be alright, as no one would be using our path to maintain that git repository.

* Improving Update Mechanism

* Improving Update Process by omitting the count parameter

* Readme file clarification

* Fixed mnode_helper script being invoked on update

* Fixed Code directory creation bug

* Moved old daemon deletion into build_mn_from_source

That way, it doesn't get lost if something went wrong earlier in the script.

* Code path creation bugfix